refactor: extract chat sidebar content into SingleChatSidebar and GroupChatSidebar components
